Refactor sources and modeling folders to contain an acset directory
