+++ mira/sources/acset/stockflow.py
@@ -3,7 +3,7 @@ import sympy
 import requests
 
 from mira.sources.util import get_sympy, transition_to_templates
-from mira.modeling.ascet.stockflow_ascet import *
+from mira.modeling.acset.stockflow import *
 
 
 def template_model_from_stockflow_ascet_json(model_json) -> TemplateModel:
